requirements.txt
Flask>=2.3.0
Pillow>=9.0.0